+++ tests/Unit/Bootstrap5/Inputs/MarginBottom/InputMarginBottomTest.php
@@ -15,13 +15,6 @@ class InputMarginBottomTest extends TestCase
     }
 
     /** @test */
-    public function it_can_enable_input_file_margin_bottom_by_default(): void
-    {
-        $html = $this->renderComponent(Input::class, ['name' => 'first_name', 'type' => 'file']);
-        self::assertStringContainsString('<div class="mb-3">', $html);
-    }
-
-    /** @test */
     public function it_can_disable_input_margin_bottom(): void
     {
         $html = $this->renderComponent(Input::class, ['name' => 'first_name', 'marginBottom' => false]);